requirements.txt
matplotlib==3.7.1
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==9.4.0
Pillow==9.5.0
scipy==1.10.1
torch==2.0.0+cu117
torchmetrics==0.11.4
torchvision==0.15.1+cu117
tqdm==4.65.0
